pyproject.toml
[project]
name = "game"
version = "0.1.0"
description = "PPO agent for IADS penetration with missiles + decoys"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0",
    "gymnasium>=1.0.0",
    "torch>=2.0",
    "tensorboard>=2.16.0",
    "onnxscript>=0.7.0",
    "onnxruntime>=1.16.0",
    "matplotlib>=3.8.0",
    "moviepy>=1.0.3,<2.0",
    "pygame-ce>=2.5.7",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
include = ["iads*", "ppo*"]

# PyTorch wheels with CUDA 12.8. Used on Linux only -- macOS pulls the standard
# wheels (CPU + MPS) from PyPI because the cu128 index has no Mac builds.
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]
